+++ contracts/ItemMock.sol
@@ -1,0 +1,172 @@
+/**
+ *  @authors: [@mtsalenc]
+ *  @reviewers: []
+ *  @auditors: []
+ *  @bounties: []
+ *  @deployments: []
+ */
+
+pragma solidity ^0.5.1;
+
+contract ItemMock {
+
+    bytes[] public items;
+    mapping(bytes32 => bool) keys;
+    Column[] columns;
+
+    struct Column {
+        uint offset;
+        uint length;
+    }
+
+
+    constructor(uint[] memory _offsets, uint[] memory _lengths) public {
+        for (uint i = 0; i < _offsets.length; i++) {
+            columns.push(Column(_offsets[i], _lengths[i]));
+        }
+    }
+
+    // ["0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x01","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x14","0x72","0x3a","0x09","0xac","0xff","0x6d","0x2a","0x60","0xdc","0xdf","0x7a","0xa4","0xaf","0xf3","0x08","0xfd","0xdc","0x16","0x0c"]
+    // ["0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x05","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0x00","0xca","0x35","0xb7","0xd9","0x15","0x45","0x8e","0xf5","0x40","0xad","0xe6","0x06","0x8d","0xfe","0x2f","0x44","0xe8","0xfa","0x73","0x3c"]
+
+    function addItem(bytes calldata _item) external {
+        bytes32 key = keccak256(_item);
+        require(!keys[key], "Item already added");
+        items.push(_item);
+        keys[key] = true;
+    }
+
+    /**
+     * @dev Search the list of items by columns.
+     * @return Any item that with a column value that matches the column provided at `_selector`.
+     */
+    function getItem(bytes calldata _selector, uint cursor, bool[] calldata _skipColumn) external view returns (uint[10] memory) {
+        uint[10] memory indexes;
+        uint count = 0;
+        for(uint i = cursor; i < items.length; i++) { // Iterate over list of items starting at cursor.
+            bytes storage item = items[i];
+            for(uint j = 0; j < columns.length; j++) { // Iterate over each column.
+                if (_skipColumn[j]) continue; // Skip that column if the caller doesn't want to provide it.
+                uint offset = columns[j].offset;
+                uint length = columns[j].length;
+                bytes memory val = slice(item, offset, length);
+                bytes memory selector = slice(_selector, offset, length);
+                if (equal(val, selector)) {
+                    indexes[count] = i;
+                    count++;
+                    break;
+                }
+            }
+        }
+        return indexes;
+    }
+
+    function equal(bytes memory _preBytes, bytes memory _postBytes) internal pure returns (bool) {
+        bool success = true;
+
+        assembly {
+            let length := mload(_preBytes)
+
+            // if lengths don't match the arrays are not equal
+            switch eq(length, mload(_postBytes))
+            case 1 {
+                // cb is a circuit breaker in the for loop since there's
+                //  no said feature for inline assembly loops
+                // cb = 1 - don't breaker
+                // cb = 0 - break
+                let cb := 1
+
+                let mc := add(_preBytes, 0x20)
+                let end := add(mc, length)
+
+                for {
+                    let cc := add(_postBytes, 0x20)
+                // the next line is the loop condition:
+                // while(uint(mc < end) + cb == 2)
+                } eq(add(lt(mc, end), cb), 2) {
+                    mc := add(mc, 0x20)
+                    cc := add(cc, 0x20)
+                } {
+                    // if any of these checks fails then arrays are not equal
+                    if iszero(eq(mload(mc), mload(cc))) {
+                        // unsuccess:
+                        success := 0
+                        cb := 0
+                    }
+                }
+            }
+            default {
+                // unsuccess:
+                success := 0
+            }
+        }
+
+        return success;
+    }
+
+    function slice(
+        bytes memory _bytes,
+        uint _start,
+        uint _length
+    )
+        internal
+        pure
+        returns (bytes memory)
+    {
+        require(_bytes.length >= (_start + _length));
+
+        bytes memory tempBytes;
+
+        assembly {
+            switch iszero(_length)
+            case 0 {
+                // Get a location of some free memory and store it in tempBytes as
+                // Solidity does for memory variables.
+                tempBytes := mload(0x40)
+
+                // The first word of the slice result is potentially a partial
+                // word read from the original array. To read it, we calculate
+                // the length of that partial word and start copying that many
+                // bytes into the array. The first word we copy will start with
+                // data we don't care about, but the last `lengthmod` bytes will
+                // land at the beginning of the contents of the new array. When
+                // we're done copying, we overwrite the full first word with
+                // the actual length of the slice.
+                let lengthmod := and(_length, 31)
+
+                // The multiplication in the next line is necessary
+                // because when slicing multiples of 32 bytes (lengthmod == 0)
+                // the following copy loop was copying the origin's length
+                // and then ending prematurely not copying everything it should.
+                let mc := add(add(tempBytes, lengthmod), mul(0x20, iszero(lengthmod)))
+                let end := add(mc, _length)
+
+                for {
+                    // The multiplication in the next line has the same exact purpose
+                    // as the one above.
+                    let cc := add(add(add(_bytes, lengthmod), mul(0x20, iszero(lengthmod))), _start)
+                } lt(mc, end) {
+                    mc := add(mc, 0x20)
+                    cc := add(cc, 0x20)
+                } {
+                    mstore(mc, mload(cc))
+                }
+
+                mstore(tempBytes, _length)
+
+                //update free-memory pointer
+                //allocating the array padded to 32 bytes like the compiler does now
+                mstore(0x40, and(add(mc, 31), not(31)))
+            }
+            //if we want a zero-length slice let's just return a zero-length array
+            default {
+                tempBytes := mload(0x40)
+
+                mstore(0x40, add(tempBytes, 0x20))
+            }
+        }
+
+        return tempBytes;
+    }
+
+}