+++ src/containers/factory/index.js
@@ -1,6 +1,5 @@
-import { Layout, Steps, Button, Icon } from 'antd'
+import { Layout, Steps, Button, Icon, Card, Empty } from 'antd'
 import React, { useState, useEffect } from 'react'
-import PropTypes from 'prop-types'
 import styled from 'styled-components/macro'
 import TCRParams from './tcr-params'
 import ItemParams from './item-params'
@@ -18,18 +17,22 @@ const StyledStepper = styled.div`
 `
 const StyledContainer = styled.div`
   margin: 32px 0;
+  word-break: break-word;
 `
 const formIds = ['tcrParamsForm', 'itemParamsForm', 'deployTCRForm']
-const CurrentStep = ({ currStep, ...rest }) => (
+const CurrentStep = props => (
   <>
     {(() => {
+      const {
+        tcrState: { currStep }
+      } = props
       switch (currStep) {
         case 1:
-          return <TCRParams formId={formIds[currStep]} {...rest} />
+          return <TCRParams formId={formIds[currStep]} {...props} />
         case 2:
-          return <ItemParams formId={formIds[currStep]} {...rest} />
+          return <ItemParams formId={formIds[currStep]} {...props} />
         case 3:
-          return <Deploy formId={formIds[currStep]} {...rest} />
+          return <Deploy formId={formIds[currStep]} {...props} />
         default:
           throw new Error('Unknown step')
       }
@@ -37,25 +40,9 @@ const CurrentStep = ({ currStep, ...rest }) => (
   </>
 )
 
-CurrentStep.propTypes = {
-  currStep: PropTypes.number.isRequired
-}
-
-const useStepper = initialStep => {
-  const [currStep, setStep] = useState(initialStep)
-  const STEP_COUNT = 3
-  const nextStep = () =>
-    setStep(currStep => (currStep === STEP_COUNT ? currStep : currStep + 1))
-  const previousStep = () =>
-    setStep(currStep => (currStep === 1 ? currStep : currStep - 1))
-  const resetStepper = () => setStep(1)
-
-  return { currStep, nextStep, previousStep, resetStepper }
-}
-
 const useCachedFactory = version => {
   const key = `tcrState@${version}`
-  const initialState = {
+  const initialWizardState = {
     title: '',
     description: '',
     requestDeposit: 0.1,
@@ -71,37 +58,71 @@ const useCachedFactory = version => {
     ],
     currStep: 1
   }
+  const initialState = {
+    ...initialWizardState,
+    transactions: {}
+  }
   let cache = window.localStorage.getItem(key)
   if (cache) cache = JSON.parse(cache)
   else cache = JSON.parse(JSON.stringify(initialState)) // Deep copy.
 
   const [tcrState, setTcrState] = useState(cache)
-  const { currStep, nextStep, previousStep } = useStepper(cache.currStep)
-  const resetTcrState = () =>
-    setTcrState(JSON.parse(JSON.stringify(initialState)))
 
-  useEffect(() =>
-    window.localStorage.setItem(
-      key,
-      JSON.stringify({
-        ...tcrState,
-        currStep
-      })
-    )
+  const STEP_COUNT = 3
+  const nextStep = () =>
+    setTcrState(prevState => ({
+      ...prevState,
+      currStep:
+        prevState.currStep === STEP_COUNT
+          ? prevState.currStep
+          : prevState.currStep + 1
+    }))
+  const previousStep = () =>
+    setTcrState(prevState => ({
+      ...prevState,
+      currStep:
+        prevState.currStep === 1 ? prevState.currStep : prevState.currStep - 1
+    }))
+  const resetStepper = () =>
+    setTcrState(prevState => ({ ...prevState, currStep: 1 }))
+  const resetTcrState = () =>
+    setTcrState(prevState => ({
+      ...JSON.parse(JSON.stringify(initialWizardState)),
+      transactions: prevState.transactions
+    }))
+  const setTxState = tx =>
+    setTcrState(prevState => ({
+      ...prevState,
+      transactions: {
+        ...prevState.transactions,
+        [tx.txHash]: tx
+      }
+    }))
+
+  useEffect(
+    () => window.localStorage.setItem(key, JSON.stringify({ ...tcrState })),
+    [tcrState, key]
   )
+
   return {
     tcrState,
     setTcrState,
     resetTcrState,
-    currStep,
     nextStep,
-    previousStep
+    previousStep,
+    resetStepper,
+    STEP_COUNT,
+    setTxState
   }
 }
 
 export default () => {
   const cachedFactory = useCachedFactory(version)
-  const { currStep, nextStep, previousStep } = cachedFactory
+  const {
+    tcrState: { currStep, transactions },
+    nextStep,
+    previousStep
+  } = cachedFactory
 
   return (
     <Content>
@@ -134,6 +155,17 @@ export default () => {
           </Button>
         </ButtonGroup>
       </StyledStepper>
+      <StyledContainer>
+        <Card title="Previous Deployments">
+          {Object.keys(transactions).length > 0 ? (
+            Object.keys(transactions).map((txHash, i) => (
+              <div key={i}>{transactions[txHash].contractAddress}</div>
+            ))
+          ) : (
+            <Empty description={false} image={Empty.PRESENTED_IMAGE_SIMPLE} />
+          )}
+        </Card>
+      </StyledContainer>
     </Content>
   )
 }